Use signed int for all loops parallelized with OpenMP
Some OpenMP versions do not support unsigned integers

+++ brian2/devices/cpp_standalone/templates/group_variable_set.cpp
@@ -8,7 +8,7 @@
     {{scalar_code|autoindent}}
 
     {{ openmp_pragma('parallel-static') }}
-	for(size_t _idx_group_idx=0; _idx_group_idx<_num_group_idx; _idx_group_idx++)
+	for(int _idx_group_idx=0; _idx_group_idx<(int)_num_group_idx; _idx_group_idx++)
 	{
 	    // vector code
 		const size_t _idx = {{_group_idx}}[_idx_group_idx];
